Rename AbstractDialog to StandardDialog, make the class concrete
